refactor: 유저-인증 엔드포인트 분리
